+++ src/jrd/err.cpp
@@ -278,7 +278,7 @@ bool ERR_post_warning(const Arg::StatusVector& v)
  **************************************/
 	fb_assert(v.value()[0] == isc_arg_warning);
 
-	int indx = 0, warning_indx = 0;
+	size_t indx = 0, warning_indx = 0;
 	ISC_STATUS* const status_vector = JRD_get_thread_data()->tdbb_status_vector;
 
 	if (status_vector[0] != isc_arg_gds ||
@@ -298,9 +298,9 @@ bool ERR_post_warning(const Arg::StatusVector& v)
 	}
 
 	/* stuff the warning */
-	if (indx + v.length() + 1 < ISC_STATUS_LENGTH)
+	if (indx + static_cast<size_t>(v.length()) + 1 < ISC_STATUS_LENGTH)
 	{
-		memcpy(&status_vector[indx], v.value(), sizeof(ISC_STATUS) * (v.length() + 1));
+		memcpy(&status_vector[indx], v.value(), sizeof(ISC_STATUS) * (static_cast<size_t>(v.length()) + 1));
         ERR_make_permanent(&status_vector[indx]);
 		return true;
 	}
@@ -395,7 +395,7 @@ static void internal_post(const ISC_STATUS* tmp_status)
  **************************************/
 
 	/* calculate length of the status */
-	int tmp_status_len = 0, warning_indx = 0;
+	size_t tmp_status_len = 0, warning_indx = 0;
 	PARSE_STATUS(tmp_status, tmp_status_len, warning_indx);
 	fb_assert(warning_indx == 0);
 
@@ -410,13 +410,13 @@ static void internal_post(const ISC_STATUS* tmp_status)
 		return;
 	}
 
-	int status_len = 0;
+	size_t status_len = 0;
 	PARSE_STATUS(status_vector, status_len, warning_indx);
 	if (status_len)
 		--status_len;
 
 	/* check for duplicated error code */
-	int i;
+	size_t i;
 	for (i = 0; i < ISC_STATUS_LENGTH; i++)
 	{
 		if (status_vector[i] == isc_arg_end && i == status_len)
@@ -435,12 +435,12 @@ static void internal_post(const ISC_STATUS* tmp_status)
 	}
 
 /* if the status_vector has only warnings then adjust err_status_len */
-	int err_status_len = i;
+	size_t err_status_len = i;
 	if (err_status_len == 2 && warning_indx)
 		err_status_len = 0;
 
 	ISC_STATUS_ARRAY warning_status;
-	int warning_count = 0;
+	size_t warning_count = 0;
 	if (warning_indx)
 	{
 		/* copy current warning(s) to a temp buffer */
